[issue-1] Option 변경에 대한 메서드구현
- 물건을 구매를 하면 해당 옵션에 대한 재고가 변경이 된다. 이 부분을 구현

+++ src/main/java/com/example/demo/repo/OptionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repo;
 import com.example.demo.domain.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OptionRepository extends JpaRepository<Option, Long> {
+public interface OptionRepository extends JpaRepository<Option, Integer> {
 }
